feat #13 : 챌린지 erd 변경 건에 따라 엔터티 생성 및 관계 주입
