--- src/main/java/com/umc/yourun/domain/enums/ChallengeKind.java
@@ -1,6 +1,0 @@
-package com.umc.yourun.domain.enums;
-
-public enum ChallengeKind {
-
-    CREW, SOLO
-}
