+++ src/main/java/com/umc/yourun/domain/enums/ChallengeStatus.java
@@ -2,5 +2,5 @@ package com.umc.yourun.domain.enums;
 
 public enum ChallengeStatus {
 
-    START, DONE, GIVEUP
+    WAIT, CHALLENGE, DONE
 }
