+++ src/main/resources/application.yml
@@ -8,7 +8,7 @@ spring:
     password: ${DB_PASSWORD}
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: none
       jdbc:
         time_zone: Asia/Seoul
   application:
